wizard_spider/Resources/Emotet_Dropper/requirements.txt
pyyaml>=5.1